dix: Fix ATOM typedef
unsigned long is needlessly large on LP64.  Use uint32_t instead.

+++ include/misc.h
@@ -79,6 +79,7 @@ OF THIS SOFTWARE.
 #include <X11/Xdefs.h>
 
 #include <stddef.h>
+#include <stdint.h>
 
 #ifndef MAXSCREENS
 #define MAXSCREENS	16
@@ -91,7 +92,7 @@ OF THIS SOFTWARE.
 #define EXTENSION_EVENT_BASE 64
 #define EXTENSION_BASE 128
 
-typedef unsigned long ATOM;
+typedef uint32_t ATOM;
 
 #ifndef TRUE
 #define TRUE 1
